feat: move boundaries to separate file
